Move inner classes from NestedLoopBatchIterator to upper level
