+++ dex/src/main/java/io/crate/data/join/LuceneLongBitSetWrapper.java
@@ -20,12 +20,13 @@
  * agreement.
  */
 
-package io.crate.data;
+package io.crate.data.join;
 
+import io.crate.data.join.NestedLoopBatchIterator;
 import org.apache.lucene.util.LongBitSet;
 
 /**
- * This BitSet is used to mark matched rows between left and right in {@link io.crate.data.NestedLoopBatchIterator}
+ * This BitSet is used to mark matched rows between left and right in {@link NestedLoopBatchIterator}
  * <p>
  * Each bit true if the rows in the respective position are matched and therefore we need
  * a structure capable of holding <pre>long</pre> size of bits so java.util.BitSet cannot be used.
