+++ sql/src/main/java/io/crate/operation/join/NestedLoopOperation.java
@@ -23,6 +23,7 @@ package io.crate.operation.join;
 
 import io.crate.concurrent.CompletionListenable;
 import io.crate.data.*;
+import io.crate.data.join.NestedLoopBatchIterator;
 import io.crate.planner.node.dql.join.JoinType;
 
 import javax.annotation.Nullable;
